Fixed a write past requested memory size

+++ include/argparsing.h
@@ -1,6 +1,8 @@
 #ifndef ARGPARSING_H_
 #define ARGPARSING_H_
 
-void parse_arguments(int argc, char **argv, char* file_name, unsigned int *input_length, unsigned int *output_length);
+#include <stddef.h>
+
+void parse_arguments(int argc, char **argv, char* file_name, size_t *input_length, size_t *output_length);
 
 #endif
